requirements.txt
PyJWT
fastapi
uvicorn
databases[postgresql]
psycopg2-binary
passlib
sqlalchemy